Load balancer should not have search domains set
Make DNS search domain only available for regular service container, not Load Balancer and external service containers.
