+++ code/iaas/service-discovery/server/src/main/java/io/cattle/platform/servicediscovery/deployment/impl/unit/ExternalDeploymentUnitInstance.java
@@ -140,4 +140,9 @@ public class ExternalDeploymentUnitInstance extends DeploymentUnitInstance {
     public boolean isIgnore() {
         return false;
     }
+
+    @Override
+    public List<String> getSearchDomains() {
+        return null;
+    }
 }
