+++ code/iaas/service-discovery/server/src/main/java/io/cattle/platform/servicediscovery/deployment/impl/unit/LoadBalancerDeploymentUnitInstance.java
@@ -73,4 +73,9 @@ public class LoadBalancerDeploymentUnitInstance extends DefaultDeploymentUnitIns
             launchConfigData.put(InstanceConstants.FIELD_PORTS, newPorts);
         }
     }
+
+    @Override
+    public List<String> getSearchDomains() {
+       return null;
+    }
 }
